Refactor: Modularize UI and standardize DB access
- I updated requirements.txt to include openpyxl and python-docx.
- I updated README.md to reflect the provided requirements.txt.
- I began modularization of main.py:
    - I moved SettingsDialog and TemplateDialog to dialogs.py.
    - I moved ClientWidget to client_widget.py.
    - I adjusted imports accordingly.
- I standardized database access in ClientWidget:
    - I replaced direct sqlite3 calls with db_manager functions for methods like load_statuses and save_client_notes.
- I confirmed i18n usage (self.tr()) in modified files.

Note: I was unable to remove the env/ and __pycache__/ directories from the git history, but .gitignore is correctly configured to ignore them for future changes.
