+++ main.py
@@ -37,6 +37,8 @@ from PyQt5.QtWidgets import QBoxLayout
 from pagedegrde import generate_cover_page_logic, APP_CONFIG as PAGEDEGRDE_APP_CONFIG # For cover page
 from docx import Document # Added for .docx support
 from company_management import CompanyTabWidget # Added for Company Management
+from dialogs import SettingsDialog, TemplateDialog # MOVED DIALOGS
+from client_widget import ClientWidget # MOVED ClientWidget
 from datetime import datetime # Ensure datetime is explicitly imported if not already for populate_docx_template
 from projectManagement import MainDashboard as ProjectManagementDashboard # Added for integration
 from html_to_pdf_util import convert_html_to_pdf # For PDF generation from HTML
@@ -289,462 +291,6 @@ class ContactDialog(QDialog):
             "position": self.position_input.text().strip(),
             "is_primary": 1 if self.primary_check.isChecked() else 0
         }
-
-class TemplateDialog(QDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Gestion des Modèles")
-        self.setMinimumSize(800, 500) # Increased size to accommodate preview
-        self.setup_ui()
-        
-    def setup_ui(self):
-        main_hbox_layout = QHBoxLayout(self) # Main layout is now QHBoxLayout
-
-        # Left side (list and buttons)
-        left_vbox_layout = QVBoxLayout()
-        left_vbox_layout.setSpacing(10) # Good.
-
-        self.template_list = QTreeWidget()
-        self.template_list.setColumnCount(4)
-        self.template_list.setHeaderLabels([self.tr("Name"), self.tr("Type"), self.tr("Language"), self.tr("Default Status")])
-
-        header = self.template_list.header() # Get header for styling
-        header.setSectionResizeMode(0, QHeaderView.Stretch)
-        header.setSectionResizeMode(1, QHeaderView.ResizeToContents)
-        header.setSectionResizeMode(2, QHeaderView.ResizeToContents)
-        header.setSectionResizeMode(3, QHeaderView.ResizeToContents)
-        self.template_list.setAlternatingRowColors(True) # Enable alternating row colors
-        
-        font = self.template_list.font()
-        font.setPointSize(font.pointSize() + 1)
-        self.template_list.setFont(font)
-        left_vbox_layout.addWidget(self.template_list)
-        
-        btn_layout = QHBoxLayout()
-        btn_layout.setSpacing(8) # Slightly tighter buttons
-
-        self.add_btn = QPushButton(self.tr("Ajouter")) # Text updated
-        self.add_btn.setIcon(QIcon.fromTheme("list-add"))
-        self.add_btn.setToolTip(self.tr("Ajouter un nouveau modèle"))
-        self.add_btn.setObjectName("primaryButton") # Style as primary
-        self.add_btn.clicked.connect(self.add_template)
-        btn_layout.addWidget(self.add_btn)
-        
-        self.edit_btn = QPushButton(self.tr("Modifier")) # Text updated
-        self.edit_btn.setIcon(QIcon.fromTheme("document-edit"))
-        self.edit_btn.setToolTip(self.tr("Modifier le modèle sélectionné (ouvre le fichier externe)"))
-        self.edit_btn.clicked.connect(self.edit_template)
-        self.edit_btn.setEnabled(False)
-        btn_layout.addWidget(self.edit_btn)
-        
-        self.delete_btn = QPushButton(self.tr("Supprimer")) # Text updated
-        self.delete_btn.setIcon(QIcon.fromTheme("edit-delete"))
-        self.delete_btn.setToolTip(self.tr("Supprimer le modèle sélectionné"))
-        self.delete_btn.setObjectName("dangerButton") # Style as danger
-        self.delete_btn.clicked.connect(self.delete_template)
-        self.delete_btn.setEnabled(False)
-        btn_layout.addWidget(self.delete_btn)
-        
-        self.default_btn = QPushButton(self.tr("Par Défaut")) # Text updated
-        self.default_btn.setIcon(QIcon.fromTheme("emblem-default")) # Alternative: "star"
-        self.default_btn.setToolTip(self.tr("Définir le modèle sélectionné comme modèle par défaut pour sa catégorie et langue"))
-        self.default_btn.clicked.connect(self.set_default_template)
-        self.default_btn.setEnabled(False)
-        btn_layout.addWidget(self.default_btn)
-        
-        left_vbox_layout.addLayout(btn_layout)
-        main_hbox_layout.addLayout(left_vbox_layout, 1)
-
-        # Right side (preview area)
-        self.preview_area = QTextEdit()
-        self.preview_area.setReadOnly(True)
-        self.preview_area.setPlaceholderText(self.tr("Sélectionnez un modèle pour afficher un aperçu."))
-        self.preview_area.setStyleSheet( # Updated stylesheet
-            "QTextEdit {"
-            "    border: 1px solid #cccccc;"
-            "    background-color: #f9f9f9;"
-            "    padding: 5px;"
-            "}"
-        )
-        main_hbox_layout.addWidget(self.preview_area, 2)
-
-        # Set overall dialog layout margins
-        main_hbox_layout.setContentsMargins(15, 15, 15, 15) # More padding
-
-        self.load_templates()
-        self.template_list.currentItemChanged.connect(self.handle_tree_item_selection) # Changed signal
-
-    def handle_tree_item_selection(self, current_item, previous_item):
-        if current_item is not None and current_item.parent() is not None: # It's a child (template) item
-            self.show_template_preview(current_item)
-            self.edit_btn.setEnabled(True)
-            self.delete_btn.setEnabled(True)
-            self.default_btn.setEnabled(True)
-        else: # It's a category item or selection cleared
-            self.preview_area.clear()
-            self.preview_area.setPlaceholderText(self.tr("Sélectionnez un modèle pour afficher un aperçu."))
-            self.edit_btn.setEnabled(False)
-            self.delete_btn.setEnabled(False)
-            self.default_btn.setEnabled(False)
-
-    def show_template_preview(self, item):
-        if not item: # Should be caught by handle_tree_item_selection, but good practice
-            self.preview_area.clear()
-            self.preview_area.setPlaceholderText(self.tr("Sélectionnez un modèle pour afficher un aperçu."))
-            return
-
-        template_id = item.data(0, Qt.UserRole) # Data is in column 0 for QTreeWidgetItems
-        if template_id is None: # It's a category item or item without template_id
-            self.preview_area.clear()
-            self.preview_area.setPlaceholderText(self.tr("Sélectionnez un modèle pour afficher un aperçu."))
-            return
-
-        # conn = None # No longer needed here
-        try:
-            # conn = sqlite3.connect(DATABASE_NAME) # Replaced
-            # cursor = conn.cursor() # Replaced
-            # # Use base_file_name and language_code as per table structure, aliasing language_code to language for consistency with other methods if needed
-            # cursor.execute("SELECT base_file_name, language_code FROM Templates WHERE template_id = ?", (template_id,)) # Replaced
-            # result = cursor.fetchone() # Replaced
-            details = db_manager.get_template_details_for_preview(template_id)
-
-            if details:
-                base_file_name = details['base_file_name']
-                language_code = details['language_code']
-                # template_file_path = os.path.join(CONFIG["templates_dir"], language_code, base_file_name) # Path construction remains the same
-
-                # The rest of the path and file handling logic remains the same
-                template_file_path = os.path.join(CONFIG["templates_dir"], language_code, base_file_name)
-                self.preview_area.clear()
-                if os.path.exists(template_file_path):
-                    _, file_extension = os.path.splitext(template_file_path)
-                    file_extension = file_extension.lower()
-
-                    if file_extension == ".xlsx":
-                        try:
-                            df = pd.read_excel(template_file_path, sheet_name=0) # Read first sheet
-                            # PREPEND CSS for basic styling
-                            html_content = f"""
-                            <style>
-                                table {{ border-collapse: collapse; width: 95%; font-family: Arial, sans-serif; margin: 10px; }} /* Adjusted width and added margin */
-                                th, td {{ border: 1px solid #cccccc; padding: 6px; text-align: left; }} /* Reduced padding slightly */
-                                th {{ background-color: #e0e0e0; font-weight: bold; }} /* Slightly darker header, bold text */
-                                td {{ text-align: right; }} /* Default text-align right for cells */
-                                tr:nth-child(even) {{ background-color: #f9f9f9; }}
-                                tr:hover {{ background-color: #e6f7ff; }} /* Add a hover effect for rows */
-                            </style>
-                            {df.to_html(escape=False, index=False, border=0)}
-                            """ # border=0 for df.to_html because CSS handles it
-                            self.preview_area.setHtml(html_content)
-                        except Exception as e:
-                            self.preview_area.setPlainText(self.tr("Erreur de lecture du fichier Excel:\n{0}").format(str(e)))
-                    elif file_extension == ".docx":
-                        try:
-                            doc = Document(template_file_path)
-                            full_text = []
-                            for para in doc.paragraphs:
-                                full_text.append(para.text)
-                            self.preview_area.setPlainText("\n".join(full_text))
-                        except Exception as e:
-                            self.preview_area.setPlainText(self.tr("Erreur de lecture du fichier Word:\n{0}").format(str(e)))
-                    elif file_extension == ".html":
-                        try:
-                            with open(template_file_path, "r", encoding="utf-8") as f:
-                                self.preview_area.setPlainText(f.read())
-                        except Exception as e:
-                            self.preview_area.setPlainText(self.tr("Erreur de lecture du fichier HTML:\n{0}").format(str(e)))
-                    else:
-                        self.preview_area.setPlainText(self.tr("Aperçu non disponible pour ce type de fichier."))
-                else:
-                    self.preview_area.setPlainText(self.tr("Fichier modèle introuvable."))
-            else:
-                self.preview_area.setPlainText(self.tr("Détails du modèle non trouvés dans la base de données."))
-
-        # except sqlite3.Error as e_db: # Replaced by db_manager's error handling (prints to console)
-            # self.preview_area.setPlainText(self.tr("Erreur DB lors de la récupération des détails du modèle:\n{0}").format(str(e_db)))
-        except Exception as e_general: # Catch any other unexpected errors, including those from db_manager if they are not caught by db_manager itself
-            self.preview_area.setPlainText(self.tr("Une erreur est survenue lors de la récupération des détails du modèle:\n{0}").format(str(e_general)))
-        # finally: # conn is no longer managed here
-            # if conn:
-                # conn.close()
-
-    # def update_preview(self, current_item, previous_item): # Original currentItemChanged connection
-    #     if not current_item:
-    #         self.preview_area.clear()
-    #         return
-    #     template_id = current_item.data(Qt.UserRole)
-    #     # TODO: Fetch template content based on ID and display it (Handled by show_template_preview now)
-    #     # This will likely involve reading the template file
-    #     # For now, just showing the ID as placeholder
-    #     # self.preview_area.setText(f"Preview for Template ID: {template_id}\n\n(Content loading not yet implemented)")
-    #     self.show_template_preview(current_item) # Call the new method if using currentItemChanged
-
-    def load_templates(self):
-        self.template_list.clear()
-        self.preview_area.clear()
-        self.preview_area.setPlaceholderText(self.tr("Sélectionnez un modèle pour afficher un aperçu."))
-
-        categories = db_manager.get_all_template_categories()
-        if categories is None: categories = []
-
-        if not categories:
-            # Optionally, add a "General" category if none exist, or inform the user.
-            # For now, we assume "General" is created by db.py's initialize_database.
-            # If still no categories, the tree will be empty.
-            # A message could be shown: self.template_list.addTopLevelItem(QTreeWidgetItem([self.tr("No categories found")]))
-            pass
-
-        for category_dict in categories:
-            category_item = QTreeWidgetItem(self.template_list, [category_dict['category_name']])
-            # Category items are not selectable for actions like edit/delete template
-            # category_item.setFlags(category_item.flags() & ~Qt.ItemIsSelectable) # This makes them non-selectable by mouse click for currentItemChanged
-            # Instead, we check item.parent() in handle_tree_item_selection
-
-            templates_in_category = db_manager.get_templates_by_category_id(category_dict['category_id'])
-            if templates_in_category is None: templates_in_category = []
-
-            for template_dict in templates_in_category:
-                template_name = template_dict['template_name']
-                template_type = template_dict.get('template_type', 'N/A')
-                language = template_dict['language_code']
-                is_default = self.tr("Yes") if template_dict.get('is_default_for_type_lang') else self.tr("No")
-
-                template_item = QTreeWidgetItem(category_item, [template_name, template_type, language, is_default])
-                template_item.setData(0, Qt.UserRole, template_dict['template_id'])
-
-        self.template_list.expandAll()
-        # Ensure buttons are correctly disabled initially as no item is selected by default
-        self.edit_btn.setEnabled(False)
-        self.delete_btn.setEnabled(False)
-        self.default_btn.setEnabled(False)
-            
-    def add_template(self):
-        file_path, _ = QFileDialog.getOpenFileName(
-            self,
-            self.tr("Sélectionner un modèle"),
-            CONFIG["templates_dir"],
-            self.tr("Fichiers Modèles (*.xlsx *.docx *.html);;Fichiers Excel (*.xlsx);;Documents Word (*.docx);;Documents HTML (*.html);;Tous les fichiers (*)")
-        )
-        if not file_path: return
-            
-        name, ok = QInputDialog.getText(self, self.tr("Nom du Modèle"), self.tr("Entrez un nom pour ce modèle:"))
-        if not ok or not name.strip(): return
-
-        # Category Selection
-        existing_categories = db_manager.get_all_template_categories()
-        if existing_categories is None: existing_categories = [] # Handle case where db query fails
-
-        category_display_list = [cat['category_name'] for cat in existing_categories]
-        create_new_option = self.tr("[Create New Category...]")
-        category_display_list.append(create_new_option)
-
-        selected_category_name, ok = QInputDialog.getItem(self, self.tr("Select Template Category"),
-                                                          self.tr("Category:"), category_display_list, 0, False)
-        if not ok: return
-
-        final_category_id = None
-        if selected_category_name == create_new_option:
-            new_category_text, ok_new = QInputDialog.getText(self, self.tr("New Category"), self.tr("Enter name for new category:"))
-            if ok_new and new_category_text.strip():
-                final_category_id = db_manager.add_template_category(new_category_text.strip())
-                if not final_category_id:
-                    QMessageBox.warning(self, self.tr("Error"), self.tr("Could not create or find category: {0}").format(new_category_text.strip()))
-                    return
-            else:
-                return # User cancelled new category creation
-        else:
-            for cat in existing_categories:
-                if cat['category_name'] == selected_category_name:
-                    final_category_id = cat['category_id']
-                    break
-            if final_category_id is None:
-                QMessageBox.critical(self, self.tr("Error"), self.tr("Selected category not found internally. Please ensure 'General' category exists or create a new one."))
-                return
-
-        # Language Selection (moved after category)
-        languages = ["fr", "en", "ar", "tr", "pt"]
-        lang, ok = QInputDialog.getItem(self, self.tr("Langue du Modèle"), self.tr("Sélectionnez la langue:"), languages, 0, False)
-        if not ok: return
-            
-        target_dir = os.path.join(CONFIG["templates_dir"], lang)
-        os.makedirs(target_dir, exist_ok=True)
-        base_file_name = os.path.basename(file_path) 
-        target_path = os.path.join(target_dir, base_file_name) 
-
-        # Determine template_type based on file extension
-        file_ext = os.path.splitext(base_file_name)[1].lower()
-        template_type_for_db = "document_other" # Default
-        if file_ext == ".xlsx":
-            template_type_for_db = "document_excel"
-        elif file_ext == ".docx":
-            template_type_for_db = "document_word"
-        elif file_ext == ".html":
-            template_type_for_db = "document_html"
-
-        template_metadata = {
-            'template_name': name.strip(), # This is the user-provided name for the template
-            'template_type': template_type_for_db,
-            'language_code': lang,
-            'base_file_name': base_file_name, # The actual file name
-            'description': f"Modèle {name.strip()} en {lang} ({base_file_name})", # Basic description
-            'category_id': final_category_id, # Use the determined category_id
-            'is_default_for_type_lang': False # User-added templates are not default by default
-            # 'raw_template_file_data': None, # Not storing file content in DB for this path
-            # 'created_by_user_id': None # Add if user system is integrated here
-        }
-        # Ensure old 'category' text field is not accidentally passed if it was in template_metadata before
-        template_metadata.pop('category', None)
-
-
-        try:
-            # First, copy the file
-            shutil.copy(file_path, target_path)
-
-            # Then, add metadata to database via db_manager
-            new_template_id = db_manager.add_template(template_metadata)
-
-            if new_template_id:
-                self.load_templates() # Refresh list
-                QMessageBox.information(self, self.tr("Succès"), self.tr("Modèle ajouté avec succès."))
-            else:
-                # db_manager.add_template would print its own errors or return None on failure.
-                # If new_template_id is None, it means there was an issue (e.g., unique constraint).
-                # The db.py function might need adjustment to differentiate between "already exists" and "other error".
-                # For now, assume if it's None, it's an error the user should know about.
-                # If file was copied but DB entry failed, consider removing the copied file.
-                if os.path.exists(target_path): # Check if file was copied before potential DB error
-                     is_error_because_exists = False
-                     # Check if a template with the same name, type, and language already exists
-                     # This requires a way to query templates, e.g., get_templates_by_type_lang_name
-                     # For simplicity, we'll assume add_template failing with None means either unique constraint or other DB error.
-                     # A more robust check would be:
-                     # existing_templates = db_manager.get_templates_by_type(template_type_for_db, language_code=lang)
-                     # if existing_templates:
-                     #    for tpl in existing_templates:
-                     #        if tpl.get('template_name') == name.strip(): # and other unique fields if necessary
-                     #            is_error_because_exists = True
-                     #            break
-                     # A direct check for unique constraint violation in db_manager.add_template would be better.
-                     # For now, providing a generic message or trying to infer.
-                     # A simple check: if a template with this name+type+lang exists, it's likely a unique conflict.
-
-                     # Simplified check:
-                     # This is not ideal as it re-queries. db_manager.add_template should ideally signal this.
-                     # For now, we'll assume if add_template returns None, it's an issue.
-                     # The user will get a generic DB error. A more specific "already exists" would need db.add_template to provide more info.
-                     QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur lors de l'enregistrement du modèle dans la base de données. Un modèle avec des attributs similaires (nom, type, langue) existe peut-être déjà, ou une autre erreur de base de données s'est produite."))
-                     # Consider removing the copied file if DB entry failed:
-                     # os.remove(target_path)
-                else: # File copy itself might have failed before DB
-                    QMessageBox.critical(self, self.tr("Erreur Fichier"), self.tr("Erreur lors de la copie du fichier modèle."))
-
-        except Exception as e: # Catch errors from shutil.copy or other unexpected issues
-            QMessageBox.critical(self, self.tr("Erreur"), self.tr("Erreur lors de l'ajout du modèle (fichier ou DB):\n{0}").format(str(e)))
-            
-    def edit_template(self): 
-        current_item = self.template_list.currentItem()
-        if not current_item or not current_item.parent(): # Check if it's a child item
-            QMessageBox.warning(self, self.tr("Sélection Requise"), self.tr("Veuillez sélectionner un modèle (et non une catégorie) à modifier."))
-            return
-        template_id = current_item.data(0, Qt.UserRole)
-        if template_id is None: return # Should not happen if parent check passed
-
-        # conn = None # Replaced
-        try:
-            # conn = sqlite3.connect(DATABASE_NAME) # Replaced
-            # cursor = conn.cursor() # Replaced
-            # cursor.execute("SELECT file_name, language FROM Templates WHERE template_id = ?", (template_id,)) # Replaced
-            # result = cursor.fetchone() # Replaced
-            path_info = db_manager.get_template_path_info(template_id)
-            if path_info:
-                # path_info contains {'file_name': 'name.xlsx', 'language': 'fr'}
-                template_file_path = os.path.join(CONFIG["templates_dir"], path_info['language'], path_info['file_name'])
-                QDesktopServices.openUrl(QUrl.fromLocalFile(template_file_path))
-            else:
-                QMessageBox.warning(self, self.tr("Erreur"), self.tr("Impossible de récupérer les informations du modèle pour l'édition."))
-        # except sqlite3.Error as e: # Replaced by db_manager's error handling
-            # QMessageBox.warning(self, self.tr("Erreur DB"), self.tr("Erreur d'accès au modèle:\n{str(e)}"))
-        except Exception as e:
-             QMessageBox.warning(self, self.tr("Erreur"), self.tr("Erreur lors de l'ouverture du modèle:\n{0}").format(str(e)))
-        # finally: # conn is no longer managed here
-            # if conn: conn.close()
-            
-    def delete_template(self):
-        current_item = self.template_list.currentItem()
-        if not current_item or not current_item.parent(): # Check if it's a child item
-            QMessageBox.warning(self, self.tr("Sélection Requise"), self.tr("Veuillez sélectionner un modèle (et non une catégorie) à supprimer."))
-            return
-        template_id = current_item.data(0, Qt.UserRole)
-        if template_id is None: return
-
-        reply = QMessageBox.question(
-            self,
-            self.tr("Confirmer Suppression"),
-            self.tr("Êtes-vous sûr de vouloir supprimer ce modèle ?"),
-            QMessageBox.Yes | QMessageBox.No,
-            QMessageBox.No
-        )
-        if reply == QMessageBox.Yes:
-            # conn = None # Replaced
-            try:
-                # conn = sqlite3.connect(DATABASE_NAME) # Replaced
-                # cursor = conn.cursor() # Replaced
-                # cursor.execute("SELECT file_name, language FROM Templates WHERE template_id = ?", (template_id,)) # Replaced
-                # result = cursor.fetchone() # Replaced
-                # cursor.execute("DELETE FROM Templates WHERE template_id = ?", (template_id,)) # Replaced
-                # conn.commit() # Replaced
-
-                file_info = db_manager.delete_template_and_get_file_info(template_id)
-
-                if file_info:
-                    file_path_to_delete = os.path.join(CONFIG["templates_dir"], file_info['language'], file_info['file_name'])
-                    if os.path.exists(file_path_to_delete):
-                        os.remove(file_path_to_delete)
-                    self.load_templates()
-                    QMessageBox.information(self, self.tr("Succès"), self.tr("Modèle supprimé avec succès."))
-                else:
-                    # db_manager.delete_template_and_get_file_info would print its own error or if template not found.
-                    QMessageBox.critical(self, self.tr("Erreur"), self.tr("Erreur de suppression du modèle. Il est possible que le modèle n'ait pas été trouvé ou qu'une erreur de base de données se soit produite."))
-            except Exception as e: 
-                QMessageBox.critical(self, self.tr("Erreur"), self.tr("Erreur de suppression du modèle:\n{0}").format(str(e)))
-            # finally: # conn is no longer managed here
-                # if conn: conn.close()
-                
-    def set_default_template(self):
-        current_item = self.template_list.currentItem()
-        if not current_item or not current_item.parent(): # Check if it's a child item
-            QMessageBox.warning(self, self.tr("Sélection Requise"), self.tr("Veuillez sélectionner un modèle (et non une catégorie) à définir par défaut."))
-            return
-        template_id = current_item.data(0, Qt.UserRole)
-        if template_id is None: return
-
-        # conn = None # Replaced
-        try:
-            # conn = sqlite3.connect(DATABASE_NAME) # Replaced
-            # cursor = conn.cursor() # Replaced
-            # cursor.execute("SELECT template_name FROM Templates WHERE template_id = ?", (template_id,)) # Logic moved to db_manager
-            # name_result = cursor.fetchone() # Logic moved
-            # if not name_result: return # Logic moved
-            # base_name = name_result[0] # Logic moved
-            
-            # cursor.execute("UPDATE Templates SET is_default_for_type_lang = 0 WHERE template_name = ?", (base_name,)) # Logic moved
-            # cursor.execute("UPDATE Templates SET is_default_for_type_lang = 1 WHERE template_id = ?", (template_id,)) # Logic moved
-            # conn.commit() # Logic moved
-
-            success = db_manager.set_default_template_by_id(template_id)
-
-            if success:
-                self.load_templates()
-                QMessageBox.information(self, self.tr("Succès"), self.tr("Modèle défini comme modèle par défaut pour sa catégorie et langue."))
-            else:
-                # db_manager.set_default_template_by_id would print its own error or if template not found.
-                 QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur de mise à jour du modèle. Le modèle n'a peut-être pas été trouvé ou une erreur de base de données s'est produite."))
-        # except sqlite3.Error as e: # Replaced by db_manager's error handling
-            # QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur de mise à jour du modèle:\n{str(e)}"))
-        except Exception as e:
-            QMessageBox.critical(self, self.tr("Erreur"), self.tr("Erreur lors de la définition du modèle par défaut:\n{0}").format(str(e)))
-        # finally: # conn is no longer managed here
-            # if conn: conn.close()
 
 class ProductDialog(QDialog):
     def __init__(self, client_id, product_data=None, parent=None):
@@ -1753,895 +1299,6 @@ class CompilePdfDialog(QDialog):
             QMessageBox.information(self, self.tr("Email envoyé"), self.tr("Le document a été envoyé avec succès."))
         except Exception as e:
             QMessageBox.critical(self, self.tr("Erreur d'envoi"), self.tr("Erreur lors de l'envoi de l'email:\n{0}").format(str(e)))
-
-class ClientWidget(QWidget):
-    def __init__(self, client_info, config, parent=None): 
-        super().__init__(parent)
-        self.client_info = client_info
-        self.config = config 
-        self.setup_ui()
-        
-    def setup_ui(self):
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(15, 15, 15, 15) 
-        layout.setSpacing(15) 
-        
-        self.header_label = QLabel(f"<h2>{self.client_info['client_name']}</h2>")
-        self.header_label.setStyleSheet("color: #2c3e50; margin-bottom: 10px;") # Added margin-bottom
-        layout.addWidget(self.header_label)
-        
-        action_layout = QHBoxLayout()
-        
-        self.create_docs_btn = QPushButton(self.tr("Créer Documents"))
-        self.create_docs_btn.setIcon(QIcon.fromTheme("document-new"))
-        self.create_docs_btn.setObjectName("primaryButton")
-        self.create_docs_btn.clicked.connect(self.open_create_docs_dialog)
-        action_layout.addWidget(self.create_docs_btn)
-        
-        self.compile_pdf_btn = QPushButton(self.tr("Compiler PDF"))
-        self.compile_pdf_btn.setIcon(QIcon.fromTheme("document-export"))
-        self.compile_pdf_btn.setProperty("primary", True) # Alternative way to mark as primary
-        self.compile_pdf_btn.clicked.connect(self.open_compile_pdf_dialog)
-        action_layout.addWidget(self.compile_pdf_btn)
-        
-        layout.addLayout(action_layout)
-        
-        status_layout = QHBoxLayout()
-        status_label = QLabel(self.tr("Statut:"))
-        status_layout.addWidget(status_label)
-        self.status_combo = QComboBox()
-        self.load_statuses() # Statuses themselves are data, but "En cours" default might need thought
-        self.status_combo.setCurrentText(self.client_info.get("status", self.tr("En cours"))) # Default status
-        self.status_combo.currentTextChanged.connect(self.update_client_status)
-        status_layout.addWidget(self.status_combo)
-        layout.addLayout(status_layout)
-        
-        self.details_layout = QFormLayout()
-        self.details_layout.setLabelAlignment(Qt.AlignRight)
-        self.details_layout.setSpacing(8) # Added spacing for details
-
-        # Store references to value labels for easy updating
-        self.detail_value_labels = {}
-
-        # Initial population
-        self.populate_details_layout() # Call a method to populate for clarity
-
-        # Add Category display
-        self.category_label = QLabel(self.tr("Catégorie:"))
-        self.category_value_label = QLabel(self.client_info.get("category", self.tr("N/A")))
-        self.details_layout.addRow(self.category_label, self.category_value_label)
-        self.detail_value_labels["category"] = self.category_value_label
-
-
-        layout.addLayout(self.details_layout)
-        
-        notes_group = QGroupBox(self.tr("Notes"))
-        notes_layout = QVBoxLayout(notes_group)
-        self.notes_edit = QTextEdit(self.client_info.get("notes", ""))
-        self.notes_edit.setPlaceholderText(self.tr("Ajoutez des notes sur ce client..."))
-        self.notes_edit.textChanged.connect(self.save_client_notes) 
-        notes_layout.addWidget(self.notes_edit)
-        layout.addWidget(notes_group)
-        
-        self.tab_widget = QTabWidget()
-        
-        docs_tab = QWidget()
-        docs_layout = QVBoxLayout(docs_tab)
-        
-        self.doc_table = QTableWidget()
-        self.doc_table.setColumnCount(5)
-        self.doc_table.setHorizontalHeaderLabels([
-            self.tr("Nom"), self.tr("Type"), self.tr("Langue"),
-            self.tr("Date"), self.tr("Actions")
-        ])
-        self.doc_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.doc_table.setSelectionBehavior(QAbstractItemView.SelectRows)
-        self.doc_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-        docs_layout.addWidget(self.doc_table)
-        
-        doc_btn_layout = QHBoxLayout()
-        
-        refresh_btn = QPushButton(self.tr("Actualiser"))
-        refresh_btn.setIcon(QIcon.fromTheme("view-refresh"))
-        refresh_btn.clicked.connect(self.populate_doc_table)
-        doc_btn_layout.addWidget(refresh_btn)
-        
-        open_btn = QPushButton(self.tr("Ouvrir"))
-        open_btn.setIcon(QIcon.fromTheme("document-open"))
-        open_btn.clicked.connect(self.open_selected_doc)
-        doc_btn_layout.addWidget(open_btn)
-        
-        delete_btn = QPushButton(self.tr("Supprimer"))
-        delete_btn.setIcon(QIcon.fromTheme("edit-delete"))
-        delete_btn.clicked.connect(self.delete_selected_doc)
-        doc_btn_layout.addWidget(delete_btn)
-        
-        docs_layout.addLayout(doc_btn_layout)
-        self.tab_widget.addTab(docs_tab, self.tr("Documents"))
-        
-        contacts_tab = QWidget()
-        contacts_layout = QVBoxLayout(contacts_tab)
-        self.contacts_list = QListWidget()
-        self.contacts_list.setAlternatingRowColors(True)
-        self.contacts_list.itemDoubleClicked.connect(self.edit_contact)
-        contacts_layout.addWidget(self.contacts_list)
-        
-        contacts_btn_layout = QHBoxLayout()
-        self.add_contact_btn = QPushButton(self.tr("➕ Ajouter")) # Standardized text
-        self.add_contact_btn.setIcon(QIcon.fromTheme("contact-new", QIcon.fromTheme("list-add")))
-        self.add_contact_btn.setToolTip(self.tr("Ajouter un nouveau contact pour ce client"))
-        self.add_contact_btn.clicked.connect(self.add_contact)
-        contacts_btn_layout.addWidget(self.add_contact_btn)
-        
-        self.edit_contact_btn = QPushButton(self.tr("✏️ Modifier")) # Standardized text
-        self.edit_contact_btn.setIcon(QIcon.fromTheme("document-edit"))
-        self.edit_contact_btn.setToolTip(self.tr("Modifier le contact sélectionné"))
-        self.edit_contact_btn.clicked.connect(self.edit_contact)
-        contacts_btn_layout.addWidget(self.edit_contact_btn)
-        
-        self.remove_contact_btn = QPushButton(self.tr("🗑️ Supprimer")) # Standardized text
-        self.remove_contact_btn.setIcon(QIcon.fromTheme("edit-delete"))
-        self.remove_contact_btn.setToolTip(self.tr("Supprimer le lien vers le contact sélectionné pour ce client"))
-        self.remove_contact_btn.setObjectName("dangerButton") # Mark as danger
-        self.remove_contact_btn.clicked.connect(self.remove_contact)
-        contacts_btn_layout.addWidget(self.remove_contact_btn)
-        
-        contacts_layout.addLayout(contacts_btn_layout)
-        self.tab_widget.addTab(contacts_tab, self.tr("Contacts"))
-
-        products_tab = QWidget()
-        products_layout = QVBoxLayout(products_tab)
-        self.products_table = QTableWidget()
-        self.products_table.setColumnCount(6) # Adjusted for ID
-        self.products_table.setHorizontalHeaderLabels([
-            self.tr("ID"), self.tr("Nom Produit"), self.tr("Description"),
-            self.tr("Qté"), self.tr("Prix Unitaire"), self.tr("Prix Total")
-        ])
-        self.products_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-        self.products_table.setSelectionBehavior(QAbstractItemView.SelectRows)
-        self.products_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.products_table.hideColumn(0) # Hide ID column by default
-        products_layout.addWidget(self.products_table)
-        
-        products_btn_layout = QHBoxLayout()
-        self.add_product_btn = QPushButton(self.tr("➕ Ajouter")) # Standardized text
-        self.add_product_btn.setIcon(QIcon.fromTheme("list-add"))
-        self.add_product_btn.setToolTip(self.tr("Ajouter un produit pour ce client/projet"))
-        self.add_product_btn.clicked.connect(self.add_product)
-        products_btn_layout.addWidget(self.add_product_btn)
-        
-        self.edit_product_btn = QPushButton(self.tr("✏️ Modifier")) # Standardized text
-        self.edit_product_btn.setIcon(QIcon.fromTheme("document-edit"))
-        self.edit_product_btn.setToolTip(self.tr("Modifier le produit sélectionné"))
-        self.edit_product_btn.clicked.connect(self.edit_product)
-        products_btn_layout.addWidget(self.edit_product_btn)
-        
-        self.remove_product_btn = QPushButton(self.tr("🗑️ Supprimer")) # Standardized text
-        self.remove_product_btn.setIcon(QIcon.fromTheme("edit-delete"))
-        self.remove_product_btn.setToolTip(self.tr("Supprimer le produit sélectionné de ce client/projet"))
-        self.remove_product_btn.setObjectName("dangerButton") # Mark as danger
-        self.remove_product_btn.clicked.connect(self.remove_product)
-        products_btn_layout.addWidget(self.remove_product_btn)
-        
-        products_layout.addLayout(products_btn_layout)
-        self.tab_widget.addTab(products_tab, self.tr("Produits"))
-        
-        layout.addWidget(self.tab_widget)
-        
-        self.populate_doc_table()
-        self.load_contacts()
-        self.load_products()
-
-    def _handle_open_pdf_action(self, file_path):
-        print(f"Action: Open PDF for {file_path}")
-        # This method will now call generate_pdf_for_document
-        # client_info is available in self.client_info
-        if not self.client_info or 'client_id' not in self.client_info:
-            QMessageBox.warning(self, self.tr("Erreur Client"), self.tr("Les informations du client ne sont pas disponibles."))
-            return
-
-        generated_pdf_path = generate_pdf_for_document(file_path, self.client_info, self)
-        if generated_pdf_path:
-            QDesktopServices.openUrl(QUrl.fromLocalFile(generated_pdf_path))
-            # Optionally, refresh doc table if the PDF is saved in a way it should appear there
-            # self.populate_doc_table()
-
-
-    def populate_details_layout(self):
-        # Clear existing rows from details_layout if any, before repopulating
-        while self.details_layout.rowCount() > 0:
-            self.details_layout.removeRow(0)
-
-        self.detail_value_labels.clear() # Clear old references
-
-        details_data_map = {
-            "project_identifier": (self.tr("ID Projet:"), self.client_info.get("project_identifier", self.tr("N/A"))),
-            "country": (self.tr("Pays:"), self.client_info.get("country", self.tr("N/A"))),
-            "city": (self.tr("Ville:"), self.client_info.get("city", self.tr("N/A"))),
-            "need": (self.tr("Besoin Principal:"), self.client_info.get("need", self.tr("N/A"))),
-            "price": (self.tr("Prix Final:"), f"{self.client_info.get('price', 0)} €"),
-            "creation_date": (self.tr("Date Création:"), self.client_info.get("creation_date", self.tr("N/A"))),
-            "base_folder_path": (self.tr("Chemin Dossier:"), f"<a href='file:///{self.client_info.get('base_folder_path','')}'>{self.client_info.get('base_folder_path','')}</a>")
-        }
-
-        for key, (label_text, value_text) in details_data_map.items():
-            label_widget = QLabel(label_text)
-            value_widget = QLabel(value_text)
-            if key == "base_folder_path": # Special handling for link
-                value_widget.setOpenExternalLinks(True)
-                value_widget.setTextInteractionFlags(Qt.TextBrowserInteraction)
-
-            self.details_layout.addRow(label_widget, value_widget)
-            self.detail_value_labels[key] = value_widget # Store reference to the value label
-
-    def refresh_display(self, new_client_info):
-        self.client_info = new_client_info # Update internal data store
-
-        # Update header
-        self.header_label.setText(f"<h2>{self.client_info.get('client_name', '')}</h2>")
-
-        # Update status combo
-        # Note: status_combo displays status name. self.client_info should have 'status' field with the name.
-        self.status_combo.setCurrentText(self.client_info.get("status", self.tr("En cours")))
-
-        # Update details in the QFormLayout
-        if hasattr(self, 'detail_value_labels'): # Check if labels were stored
-            self.detail_value_labels["project_identifier"].setText(self.client_info.get("project_identifier", self.tr("N/A")))
-            self.detail_value_labels["country"].setText(self.client_info.get("country", self.tr("N/A")))
-            self.detail_value_labels["city"].setText(self.client_info.get("city", self.tr("N/A")))
-            self.detail_value_labels["need"].setText(self.client_info.get("need", self.tr("N/A"))) # 'need' is 'primary_need_description' in some contexts
-            self.detail_value_labels["price"].setText(f"{self.client_info.get('price', 0)} €")
-            self.detail_value_labels["creation_date"].setText(self.client_info.get("creation_date", self.tr("N/A")))
-
-            folder_path = self.client_info.get('base_folder_path','')
-            self.detail_value_labels["base_folder_path"].setText(f"<a href='file:///{folder_path}'>{folder_path}</a>")
-
-            # Update Category if the label exists
-            if "category" in self.detail_value_labels:
-                 self.detail_value_labels["category"].setText(self.client_info.get("category", self.tr("N/A")))
-            elif hasattr(self, 'category_value_label'): # Fallback for direct attribute if detail_value_labels not fully populated yet
-                 self.category_value_label.setText(self.client_info.get("category", self.tr("N/A")))
-
-        else: # Fallback or if you choose to repopulate the whole layout
-            self.populate_details_layout()
-            # Explicitly update category here too if populate_details_layout doesn't handle it fully initially
-            if hasattr(self, 'category_value_label'):
-                 self.category_value_label.setText(self.client_info.get("category", self.tr("N/A")))
-
-
-        # Update Notes
-        self.notes_edit.setText(self.client_info.get("notes", ""))
-
-        # Other elements like doc_table, contacts_list, products_table are refreshed by their own mechanisms typically.
-
-    def load_statuses(self):
-        conn = None
-        try:
-            conn = sqlite3.connect(DATABASE_NAME)
-            cursor = conn.cursor()
-            cursor.execute("SELECT status_name FROM StatusSettings")
-            for status_row in cursor.fetchall(): 
-                # Potentially translate status_row[0] if they are meant to be translated in UI
-                # For now, assuming status names from DB are either keys or already in target language
-                self.status_combo.addItem(status_row[0])
-        except sqlite3.Error as e:
-            QMessageBox.warning(self, self.tr("Erreur DB"), self.tr("Erreur de chargement des statuts:\n{0}").format(str(e)))
-        finally:
-            if conn: conn.close()
-            
-    def update_client_status(self, status_text): 
-        try:
-            status_setting = db_manager.get_status_setting_by_name(status_text, 'Client')
-            if status_setting and status_setting.get('status_id') is not None:
-                status_id_to_set = status_setting['status_id']
-                client_id_to_update = self.client_info["client_id"]
-
-                if db_manager.update_client(client_id_to_update, {'status_id': status_id_to_set}):
-                    self.client_info["status"] = status_text # Keep display name
-                    self.client_info["status_id"] = status_id_to_set # Update the id in the local map
-                    # Find the item in the main list and update its UserRole for the delegate
-                    # This part is tricky as ClientWidget doesn't have direct access to DocumentManager.client_list_widget
-                    # For now, we'll rely on a full refresh from DocumentManager if the list display needs immediate color change,
-                    # or accept that the color might only update on next full load/filter.
-                    # Consider emitting a signal if immediate update of list widget item is needed.
-                    print(f"Client {client_id_to_update} status_id updated to {status_id_to_set} ({status_text})")
-                else:
-                    QMessageBox.warning(self, self.tr("Erreur DB"), self.tr("Échec de la mise à jour du statut du client dans la DB."))
-            else:
-                QMessageBox.warning(self, self.tr("Erreur Configuration"), self.tr("Statut '{0}' non trouvé ou invalide. Impossible de mettre à jour.").format(status_text))
-        except Exception as e:
-            QMessageBox.critical(self, self.tr("Erreur Inattendue"), self.tr("Erreur de mise à jour du statut:\n{0}").format(str(e)))
-            
-    def save_client_notes(self): 
-        notes = self.notes_edit.toPlainText()
-        conn = None
-        try:
-            conn = sqlite3.connect(DATABASE_NAME)
-            cursor = conn.cursor()
-            cursor.execute("UPDATE Clients SET notes = ? WHERE client_id = ?", (notes, self.client_info["client_id"]))
-            conn.commit()
-            self.client_info["notes"] = notes 
-        except sqlite3.Error as e:
-            QMessageBox.warning(self, self.tr("Erreur DB"), self.tr("Erreur de sauvegarde des notes:\n{0}").format(str(e)))
-        finally:
-            if conn: conn.close()
-            
-    def populate_doc_table(self):
-        self.doc_table.setRowCount(0)
-        client_path = self.client_info["base_folder_path"] 
-        if not os.path.exists(client_path):
-            return
-            
-        row = 0
-        for lang in self.client_info.get("selected_languages", ["fr"]):
-            lang_dir = os.path.join(client_path, lang)
-            if not os.path.exists(lang_dir):
-                continue
-                
-            for file_name in os.listdir(lang_dir):
-                if file_name.endswith(('.xlsx', '.pdf', '.docx', '.html')):
-                    file_path = os.path.join(lang_dir, file_name)
-                    name_item = QTableWidgetItem(file_name)
-                    name_item.setData(Qt.UserRole, file_path)
-
-                    file_type_str = ""
-                    if file_name.lower().endswith('.xlsx'):
-                        file_type_str = self.tr("Excel")
-                    elif file_name.lower().endswith('.docx'):
-                        file_type_str = self.tr("Word")
-                    elif file_name.lower().endswith('.html'):
-                        file_type_str = self.tr("HTML")
-                    else:
-                        file_type_str = self.tr("PDF")
-                    mod_time = datetime.fromtimestamp(os.path.getmtime(file_path)).strftime('%Y-%m-%d %H:%M') # Date format maybe locale specific
-                    
-                    self.doc_table.insertRow(row)
-                    self.doc_table.setItem(row, 0, name_item)
-                    self.doc_table.setItem(row, 1, QTableWidgetItem(file_type_str))
-                    self.doc_table.setItem(row, 2, QTableWidgetItem(lang)) # lang code, not UI text
-                    self.doc_table.setItem(row, 3, QTableWidgetItem(mod_time))
-                    
-                    action_widget = QWidget()
-                    action_layout = QHBoxLayout(action_widget)
-                    action_layout.setContentsMargins(2, 2, 2, 2) # Small margins
-                    action_layout.setSpacing(5) # Spacing between buttons
-
-                    # Button 1: Ouvrir PDF (Placeholder)
-                    pdf_btn = QPushButton("PDF") # Placeholder icon/text
-                    pdf_btn.setIcon(QIcon.fromTheme("application-pdf", QIcon("📄"))) # Fallback icon
-                    pdf_btn.setToolTip(self.tr("Ouvrir PDF du document"))
-                    pdf_btn.setFixedSize(30, 30)
-                    pdf_btn.clicked.connect(lambda _, p=file_path: self._handle_open_pdf_action(p))
-                    action_layout.addWidget(pdf_btn)
-
-                    # Button 2: Afficher Source
-                    source_btn = QPushButton("👁️") # Placeholder icon/text
-                    source_btn.setIcon(QIcon.fromTheme("document-properties", QIcon("👁️"))) # Fallback icon
-                    source_btn.setToolTip(self.tr("Afficher le fichier source"))
-                    source_btn.setFixedSize(30, 30)
-                    source_btn.clicked.connect(lambda _, p=file_path: QDesktopServices.openUrl(QUrl.fromLocalFile(p)))
-                    action_layout.addWidget(source_btn)
-                    
-                    # Button 3: Modifier Contenu (only for Excel/HTML)
-                    if file_name.lower().endswith(('.xlsx', '.html')):
-                        edit_btn = QPushButton("✏️")
-                        edit_btn.setIcon(QIcon.fromTheme("document-edit", QIcon("✏️"))) # Fallback icon
-                        edit_btn.setToolTip(self.tr("Modifier le contenu du document"))
-                        edit_btn.setFixedSize(30, 30)
-                        edit_btn.clicked.connect(lambda _, p=file_path: self.open_document(p)) # open_document handles editor logic
-                        action_layout.addWidget(edit_btn)
-                    else:
-                        # Add a spacer or disabled button if not editable to maintain layout consistency
-                        spacer_widget = QWidget()
-                        spacer_widget.setFixedSize(30,30)
-                        action_layout.addWidget(spacer_widget)
-
-
-                    # Button 4: Supprimer
-                    delete_btn = QPushButton("🗑️")
-                    delete_btn.setIcon(QIcon.fromTheme("edit-delete", QIcon("🗑️"))) # Fallback icon
-                    delete_btn.setToolTip(self.tr("Supprimer le document"))
-                    delete_btn.setFixedSize(30, 30)
-                    delete_btn.clicked.connect(lambda _, p=file_path: self.delete_document(p))
-                    action_layout.addWidget(delete_btn)
-                    
-                    action_layout.addStretch() # Push buttons to the left if there's space
-                    action_widget.setLayout(action_layout) # Set the layout on the widget
-                    self.doc_table.setCellWidget(row, 4, action_widget)
-                    row += 1
-
-    def open_create_docs_dialog(self):
-        dialog = CreateDocumentDialog(self.client_info, self.config, self)
-        if dialog.exec_() == QDialog.Accepted:
-            self.populate_doc_table()
-            
-    def open_compile_pdf_dialog(self):
-        dialog = CompilePdfDialog(self.client_info, self)
-        dialog.exec_()
-            
-    def open_selected_doc(self):
-        selected_row = self.doc_table.currentRow()
-        if selected_row >= 0:
-            file_path_item = self.doc_table.item(selected_row, 0) # Name item stores path in UserRole
-            if file_path_item:
-                file_path = file_path_item.data(Qt.UserRole)
-                if file_path and os.path.exists(file_path):
-                    self.open_document(file_path)
-                
-    def delete_selected_doc(self):
-        selected_row = self.doc_table.currentRow()
-        if selected_row >= 0:
-            file_path_item = self.doc_table.item(selected_row, 0) # Name item stores path
-            if file_path_item:
-                file_path = file_path_item.data(Qt.UserRole)
-                if file_path and os.path.exists(file_path):
-                    self.delete_document(file_path)
-                
-    def open_document(self, file_path):
-        if os.path.exists(file_path):
-            try:
-                # Data passed to editors is data, not UI text.
-                editor_client_data = {
-                    "client_id": self.client_info.get("client_id"), # Crucial for DB context
-                    "Nom du client": self.client_info.get("client_name", ""), # Legacy key, keep for compatibility
-                    "client_name": self.client_info.get("client_name", ""), # Preferred key
-                    "company_name": self.client_info.get("company_name", ""),
-                    "Besoin": self.client_info.get("need", ""), # 'need' is primary_need_description in client_info map
-                    "primary_need_description": self.client_info.get("need", ""), # Explicitly map
-                    "project_identifier": self.client_info.get("project_identifier", ""),
-                    "country": self.client_info.get("country", ""), # This is country_name
-                    "country_id": self.client_info.get("country_id"),
-                    "city": self.client_info.get("city", ""), # This is city_name
-                    "city_id": self.client_info.get("city_id"),
-                    "price": self.client_info.get("price", 0), # raw_price
-                    "status": self.client_info.get("status"), # status name
-                    "status_id": self.client_info.get("status_id"),
-                    "selected_languages": self.client_info.get("selected_languages"), # list
-                    "notes": self.client_info.get("notes"),
-                    "creation_date": self.client_info.get("creation_date"),
-                    "category": self.client_info.get("category"),
-                    "base_folder_path": self.client_info.get("base_folder_path")
-                    # Add any other direct fields from client_info that get_document_context_data might use
-                    # or that HtmlEditor might directly use for its own templating logic.
-                }
-                if file_path.lower().endswith('.xlsx'):
-                    editor = ExcelEditor(file_path, parent=self)
-                    if editor.exec_() == QDialog.Accepted:
-                        expected_pdf_basename = os.path.splitext(os.path.basename(file_path))[0] + "_" + datetime.now().strftime('%Y%m%d') + ".pdf"
-                        expected_pdf_path = os.path.join(os.path.dirname(file_path), expected_pdf_basename)
-                        if os.path.exists(expected_pdf_path):
-                            archive_timestamp = datetime.now().strftime('%Y%m%d%H%M%S')
-                            archive_pdf_name = os.path.splitext(expected_pdf_path)[0] + f"_archive_{archive_timestamp}.pdf"
-                            try:
-                                os.rename(expected_pdf_path, os.path.join(os.path.dirname(expected_pdf_path), archive_pdf_name))
-                                print(f"Archived existing PDF to: {archive_pdf_name}")
-                            except OSError as e_archive:
-                                print(f"Error archiving PDF {expected_pdf_path}: {e_archive}")
-                        generate_pdf_for_document(file_path, self.client_info, self) # This will show info message for XLSX
-                    self.populate_doc_table()
-                elif file_path.lower().endswith('.html'):
-                    html_editor_dialog = HtmlEditor(file_path, client_data=editor_client_data, parent=self)
-                    if html_editor_dialog.exec_() == QDialog.Accepted:
-                        expected_pdf_basename = os.path.splitext(os.path.basename(file_path))[0] + "_" + datetime.now().strftime('%Y%m%d') + ".pdf"
-                        expected_pdf_path = os.path.join(os.path.dirname(file_path), expected_pdf_basename)
-                        if os.path.exists(expected_pdf_path):
-                            archive_timestamp = datetime.now().strftime('%Y%m%d%H%M%S')
-                            archive_pdf_name = os.path.splitext(expected_pdf_path)[0] + f"_archive_{archive_timestamp}.pdf"
-                            try:
-                                os.rename(expected_pdf_path, os.path.join(os.path.dirname(expected_pdf_path), archive_pdf_name))
-                                print(f"Archived existing PDF to: {archive_pdf_name}")
-                            except OSError as e_archive:
-                                print(f"Error archiving PDF {expected_pdf_path}: {e_archive}")
-                        generated_pdf_path = generate_pdf_for_document(file_path, self.client_info, self)
-                        if generated_pdf_path:
-                             print(f"Updated PDF generated at: {generated_pdf_path}")
-                    self.populate_doc_table()
-                elif file_path.lower().endswith(('.docx', '.pdf')):
-                    QDesktopServices.openUrl(QUrl.fromLocalFile(file_path))
-                else:
-                    QDesktopServices.openUrl(QUrl.fromLocalFile(file_path))
-            except Exception as e:
-                QMessageBox.warning(self, self.tr("Erreur Ouverture Fichier"), self.tr("Impossible d'ouvrir le fichier:\n{0}").format(str(e)))
-        else:
-            QMessageBox.warning(self, self.tr("Fichier Introuvable"), self.tr("Le fichier n'existe plus."))
-            self.populate_doc_table()
-            
-    def delete_document(self, file_path):
-        if not os.path.exists(file_path):
-            return
-            
-        reply = QMessageBox.question(
-            self, 
-            self.tr("Confirmer la suppression"),
-            self.tr("Êtes-vous sûr de vouloir supprimer le fichier {0} ?").format(os.path.basename(file_path)),
-            QMessageBox.Yes | QMessageBox.No,
-            QMessageBox.No
-        )
-        
-        if reply == QMessageBox.Yes:
-            try:
-                os.remove(file_path)
-                self.populate_doc_table()
-                QMessageBox.information(self, self.tr("Fichier supprimé"), self.tr("Le fichier a été supprimé avec succès."))
-            except Exception as e:
-                QMessageBox.warning(self, self.tr("Erreur"), self.tr("Impossible de supprimer le fichier:\n{0}").format(str(e)))
-    
-    def load_contacts(self):
-        self.contacts_list.clear()
-        client_uuid = self.client_info.get("client_id")
-        if not client_uuid: return
-
-        try:
-            # db_manager.get_contacts_for_client returns list of dicts with contact details
-            # and ClientContacts fields like is_primary_for_client, client_contact_id
-            contacts = db_manager.get_contacts_for_client(client_uuid)
-            if contacts is None: contacts = []
-
-            for contact in contacts:
-                contact_text = f"{contact.get('name', 'N/A')}"
-                if contact.get('phone'): contact_text += f" ({contact.get('phone')})"
-                if contact.get('is_primary_for_client'): contact_text += f" [{self.tr('Principal')}]"
-
-                item = QListWidgetItem(contact_text)
-                # Store both contact_id (from Contacts table) and client_contact_id (from ClientContacts link table)
-                item.setData(Qt.UserRole, {'contact_id': contact.get('contact_id'),
-                                           'client_contact_id': contact.get('client_contact_id'),
-                                           'is_primary': contact.get('is_primary_for_client')})
-                self.contacts_list.addItem(item)
-        except Exception as e:
-            QMessageBox.warning(self, self.tr("Erreur DB"), self.tr("Erreur de chargement des contacts:\n{0}").format(str(e)))
-            
-    def add_contact(self):
-        client_uuid = self.client_info.get("client_id")
-        if not client_uuid: return
-
-        dialog = ContactDialog(client_uuid, parent=self) # Pass client_id for context if needed by dialog
-        if dialog.exec_() == QDialog.Accepted:
-            contact_form_data = dialog.get_data() # name, email, phone, position, is_primary
-
-            try:
-                # Step 1: Add or get existing contact from global Contacts table
-                # Check if contact with this email already exists
-                existing_contact = db_manager.get_contact_by_email(contact_form_data['email'])
-                contact_id_to_link = None
-
-                if existing_contact:
-                    contact_id_to_link = existing_contact['contact_id']
-                    # Optionally, update existing contact's details if they differ (name, phone, position)
-                    update_data = {k: v for k, v in contact_form_data.items() if k in ['name', 'phone', 'position'] and v != existing_contact.get(k)}
-                    if update_data : db_manager.update_contact(contact_id_to_link, update_data)
-                else:
-                    new_contact_id = db_manager.add_contact({
-                        'name': contact_form_data['name'], 'email': contact_form_data['email'],
-                        'phone': contact_form_data['phone'], 'position': contact_form_data['position']
-                    })
-                    if new_contact_id:
-                        contact_id_to_link = new_contact_id
-                    else:
-                        QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Impossible de créer le nouveau contact global."))
-                        return
-
-                # Step 2: Link contact to client in ClientContacts
-                if contact_id_to_link:
-                    if contact_form_data['is_primary']:
-                        # Unset other primary contacts for this client
-                        client_contacts = db_manager.get_contacts_for_client(client_uuid)
-                        if client_contacts:
-                            for cc in client_contacts:
-                                if cc['is_primary_for_client']:
-                                    db_manager.update_client_contact_link(cc['client_contact_id'], {'is_primary_for_client': False})
-
-                    link_id = db_manager.link_contact_to_client(
-                        client_uuid, contact_id_to_link,
-                        is_primary=contact_form_data['is_primary']
-                    )
-                    if link_id:
-                        self.load_contacts()
-                    else:
-                        QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Impossible de lier le contact au client. Le lien existe peut-être déjà."))
-            except Exception as e:
-                QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur d'ajout du contact:\n{0}").format(str(e)))
-                
-    def edit_contact(self):
-        item = self.contacts_list.currentItem()
-        if not item: return
-
-        item_data = item.data(Qt.UserRole)
-        contact_id = item_data.get('contact_id')
-        client_contact_id = item_data.get('client_contact_id') # ID of the link
-        # is_currently_primary = item_data.get('is_primary')
-
-        client_uuid = self.client_info.get("client_id")
-        if not contact_id or not client_uuid: return
-
-        try:
-            contact_details = db_manager.get_contact_by_id(contact_id)
-            if contact_details:
-                # Need to also fetch is_primary status for this specific client-contact link
-                current_link_info = None
-                client_contacts_for_client = db_manager.get_contacts_for_client(client_uuid)
-                if client_contacts_for_client:
-                    for cc_link in client_contacts_for_client:
-                        if cc_link['contact_id'] == contact_id:
-                            current_link_info = cc_link
-                            break
-
-                is_primary_for_this_client = current_link_info['is_primary_for_client'] if current_link_info else False
-
-                dialog_data = {
-                    "name": contact_details.get('name'), "email": contact_details.get('email'),
-                    "phone": contact_details.get('phone'), "position": contact_details.get('position'),
-                    "is_primary": is_primary_for_this_client
-                }
-
-                dialog = ContactDialog(client_uuid, dialog_data, parent=self)
-                if dialog.exec_() == QDialog.Accepted:
-                    new_form_data = dialog.get_data()
-
-                    # Update global contact details
-                    db_manager.update_contact(contact_id, {
-                        'name': new_form_data['name'], 'email': new_form_data['email'],
-                        'phone': new_form_data['phone'], 'position': new_form_data['position']
-                    })
-
-                    # Update is_primary status for this client link
-                    if new_form_data['is_primary'] and not is_primary_for_this_client:
-                        # Unset other primary contacts for this client
-                        if client_contacts_for_client: # re-fetch or use from above
-                            for cc in client_contacts_for_client:
-                                if cc['contact_id'] != contact_id and cc['is_primary_for_client']:
-                                    db_manager.update_client_contact_link(cc['client_contact_id'], {'is_primary_for_client': False})
-                        db_manager.update_client_contact_link(client_contact_id, {'is_primary_for_client': True})
-                    elif not new_form_data['is_primary'] and is_primary_for_this_client:
-                        db_manager.update_client_contact_link(client_contact_id, {'is_primary_for_client': False})
-
-                    self.load_contacts()
-        except Exception as e:
-            QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur de modification du contact:\n{0}").format(str(e)))
-            
-    def remove_contact(self): # This should unlink contact from client. Optionally delete global contact if not linked elsewhere.
-        item = self.contacts_list.currentItem()
-        if not item: return
-
-        item_data = item.data(Qt.UserRole)
-        contact_id = item_data.get('contact_id')
-        client_contact_id = item_data.get('client_contact_id')
-        client_uuid = self.client_info.get("client_id")
-
-        if not client_contact_id or not client_uuid or not contact_id: return
-
-        contact_name = db_manager.get_contact_by_id(contact_id)['name'] if db_manager.get_contact_by_id(contact_id) else "Inconnu"
-
-        reply = QMessageBox.question(self, self.tr("Confirmer Suppression Lien"),
-                                     self.tr("Êtes-vous sûr de vouloir supprimer le lien vers ce contact ({0}) pour ce client ?\nLe contact global ne sera pas supprimé.").format(contact_name),
-                                     QMessageBox.Yes | QMessageBox.No, QMessageBox.No)
-        if reply == QMessageBox.Yes:
-            try:
-                unlinked = db_manager.unlink_contact_from_client(client_uuid, contact_id) # Uses client_id and contact_id
-                if unlinked:
-                    self.load_contacts()
-                else:
-                    QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur de suppression du lien contact-client."))
-            except Exception as e:
-                QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur de suppression du lien contact:\n{0}").format(str(e)))
-                
-    def add_product(self):
-        client_uuid = self.client_info.get("client_id")
-        if not client_uuid: return
-
-        dialog = ProductDialog(client_uuid, parent=self) # ProductDialog now supports multi-line
-        if dialog.exec_() == QDialog.Accepted:
-            products_list_data = dialog.get_data() # This is now a list of product dicts
-
-            # Process list of products from dialog
-            products_added_count = 0
-            for product_item_data in products_list_data: # Use a different loop variable name
-                try:
-                    # Step 1: Find or Create Global Product
-                    global_product = db_manager.get_product_by_name(product_item_data['name'])
-                    global_product_id = None
-                    current_base_unit_price = None # For checking if override is needed
-
-                    if global_product:
-                        global_product_id = global_product['product_id']
-                        current_base_unit_price = global_product.get('base_unit_price')
-                    else:
-                        new_global_product_id = db_manager.add_product({
-                            'product_name': product_item_data['name'],
-                            'description': product_item_data['description'],
-                            'base_unit_price': product_item_data['unit_price']
-                        })
-                        if new_global_product_id:
-                            global_product_id = new_global_product_id
-                            current_base_unit_price = product_item_data['unit_price']
-                        else:
-                            QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Impossible de créer le produit global '{0}'.").format(product_item_data['name']))
-                            continue # Skip to next product
-
-                    # Step 2: Link Product to Client
-                    if global_product_id:
-                        unit_price_override_val = None
-                        if current_base_unit_price is None or product_item_data['unit_price'] != current_base_unit_price:
-                            unit_price_override_val = product_item_data['unit_price']
-
-                        link_data = {
-                            'client_id': client_uuid,
-                            'project_id': None,
-                            'product_id': global_product_id,
-                            'quantity': product_item_data['quantity'],
-                            'unit_price_override': unit_price_override_val
-                        }
-                        cpp_id = db_manager.add_product_to_client_or_project(link_data)
-                        if cpp_id:
-                            products_added_count +=1
-                        else:
-                            QMessageBox.warning(self, self.tr("Erreur DB"), self.tr("Impossible de lier le produit '{0}' au client.").format(product_item_data['name']))
-                except Exception as e:
-                    QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur d'ajout du produit '{0}':\n{1}").format(product_item_data.get('name', 'Inconnu'), str(e)))
-
-            if products_added_count > 0:
-                self.load_products() # Refresh product list in ClientWidget only once after processing all
-            if products_added_count < len(products_list_data) and len(products_list_data) > 0 : # Avoid message if list was empty
-                 QMessageBox.information(self, self.tr("Information"), self.tr("Certains produits n'ont pas pu être ajoutés. Veuillez vérifier les messages d'erreur."))
-
-    def edit_product(self):
-        selected_row = self.products_table.currentRow()
-        if selected_row < 0:
-            QMessageBox.information(self, self.tr("Sélection Requise"), self.tr("Veuillez sélectionner un produit à modifier."))
-            return
-
-        cpp_id_item = self.products_table.item(selected_row, 0) # Hidden column with client_project_product_id
-        if not cpp_id_item:
-            QMessageBox.critical(self, self.tr("Erreur Données"), self.tr("ID du produit lié introuvable dans la table."))
-            return
-        client_project_product_id = cpp_id_item.data(Qt.UserRole)
-
-        try:
-            # Fetch complete details of the linked product, including global product info
-            linked_product_details = db_manager.get_client_project_product_by_id(client_project_product_id)
-
-            if not linked_product_details:
-                QMessageBox.warning(self, self.tr("Erreur"), self.tr("Détails du produit lié introuvables dans la base de données."))
-                return
-
-            # Prepare data for the EditProductLineDialog
-            # The dialog expects 'name', 'description', 'quantity', 'unit_price', 'product_id', 'client_project_product_id'
-            effective_unit_price = linked_product_details.get('unit_price_override', linked_product_details.get('base_unit_price', 0.0))
-
-            dialog_data_for_edit = {
-                "name": linked_product_details.get('product_name', ''),
-                "description": linked_product_details.get('product_description', ''),
-                "quantity": linked_product_details.get('quantity', 1.0),
-                "unit_price": effective_unit_price, # This is the price shown and edited in the dialog
-                "product_id": linked_product_details.get('product_id'),
-                "client_project_product_id": client_project_product_id,
-                # Store original base price for comparison later
-                "original_base_unit_price": linked_product_details.get('base_unit_price', 0.0)
-            }
-
-            dialog = EditProductLineDialog(product_data=dialog_data_for_edit, parent=self)
-            if dialog.exec_() == QDialog.Accepted:
-                updated_data_from_dialog = dialog.get_data()
-
-                # 1. Update Global Product (Products table) if name, description, or base price logic dictates
-                global_product_update_payload = {}
-                if updated_data_from_dialog['name'] != linked_product_details.get('product_name'):
-                    global_product_update_payload['product_name'] = updated_data_from_dialog['name']
-                if updated_data_from_dialog['description'] != linked_product_details.get('product_description'):
-                    global_product_update_payload['description'] = updated_data_from_dialog['description']
-
-                # Logic for updating base_unit_price:
-                # If name or description changed, consider the edited price as the new base price for this (potentially now distinct) product.
-                if global_product_update_payload: # If name or description changed
-                     global_product_update_payload['base_unit_price'] = updated_data_from_dialog['unit_price']
-
-                if global_product_update_payload:
-                    db_manager.update_product(updated_data_from_dialog['product_id'], global_product_update_payload)
-
-                # 2. Update Linked Product (ClientProjectProducts table)
-                # Determine unit_price_override based on the potentially updated global product's base_unit_price
-                current_global_product_info = db_manager.get_product_by_id(updated_data_from_dialog['product_id'])
-                current_global_base_price = current_global_product_info.get('base_unit_price', 0.0) if current_global_product_info else 0.0
-
-                unit_price_override_val = None
-                if float(updated_data_from_dialog['unit_price']) != float(current_global_base_price): # Compare as float
-                    unit_price_override_val = updated_data_from_dialog['unit_price']
-
-                link_update_payload = {
-                    'quantity': updated_data_from_dialog['quantity'],
-                    'unit_price_override': unit_price_override_val
-                    # total_price_calculated will be handled by db_manager.update_client_project_product
-                }
-
-                if db_manager.update_client_project_product(client_project_product_id, link_update_payload):
-                    self.load_products() # Refresh the table
-                    QMessageBox.information(self, self.tr("Succès"), self.tr("Ligne de produit mise à jour avec succès."))
-                else:
-                    QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Échec de la mise à jour de la ligne de produit liée."))
-            # Else (dialog cancelled), do nothing
-        except Exception as e:
-            QMessageBox.critical(self, self.tr("Erreur Inattendue"), self.tr("Erreur lors de la modification du produit:\n{0}").format(str(e)))
-            print(f"Error in edit_product: {e}") # For debugging
-
-    def remove_product(self):
-        selected_row = self.products_table.currentRow()
-        if selected_row < 0: return
-        
-        cpp_id_item = self.products_table.item(selected_row, 0)
-        if not cpp_id_item: return
-        client_project_product_id = cpp_id_item.data(Qt.UserRole) # This is client_project_product_id
-
-        product_name_item = self.products_table.item(selected_row, 1)
-        product_name = product_name_item.text() if product_name_item else self.tr("Inconnu")
-        
-        reply = QMessageBox.question(
-            self, 
-            self.tr("Confirmer Suppression"),
-            self.tr("Êtes-vous sûr de vouloir supprimer le produit '{0}' de ce client/projet?").format(product_name),
-            QMessageBox.Yes | QMessageBox.No, QMessageBox.No
-        )
-        
-        if reply == QMessageBox.Yes:
-            try:
-                if db_manager.remove_product_from_client_or_project(client_project_product_id):
-                    self.load_products()
-                else:
-                    QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur de suppression du produit lié."))
-            except Exception as e:
-                QMessageBox.critical(self, self.tr("Erreur DB"), self.tr("Erreur de suppression du produit lié:\n{0}").format(str(e)))
-
-    def load_products(self):
-        self.products_table.setRowCount(0)
-        client_uuid = self.client_info.get("client_id")
-        if not client_uuid: return
-
-        try:
-            # Get products linked to this client (project_id=None for client-general products)
-            # This returns a list of dicts, with joined info from Products table
-            linked_products = db_manager.get_products_for_client_or_project(client_uuid, project_id=None)
-            if linked_products is None: linked_products = []
-            
-            for row_idx, prod_link_data in enumerate(linked_products):
-                self.products_table.insertRow(row_idx)
-                
-                # Store client_project_product_id for edit/delete
-                id_item = QTableWidgetItem(str(prod_link_data.get('client_project_product_id')))
-                id_item.setData(Qt.UserRole, prod_link_data.get('client_project_product_id'))
-                self.products_table.setItem(row_idx, 0, id_item)
-                
-                self.products_table.setItem(row_idx, 1, QTableWidgetItem(prod_link_data.get('product_name', 'N/A')))
-                self.products_table.setItem(row_idx, 2, QTableWidgetItem(prod_link_data.get('product_description', '')))
-                
-                qty_item = QTableWidgetItem(str(prod_link_data.get('quantity', 0)))
-                qty_item.setTextAlignment(Qt.AlignRight | Qt.AlignVCenter)
-                self.products_table.setItem(row_idx, 3, qty_item)
-                
-                unit_price_override = prod_link_data.get('unit_price_override')
-                base_price = prod_link_data.get('base_unit_price')
-
-                if unit_price_override is not None:
-                    effective_unit_price = unit_price_override
-                elif base_price is not None:
-                    effective_unit_price = base_price
-                else:
-                    effective_unit_price = 0.0
-
-                effective_unit_price = float(effective_unit_price) if effective_unit_price is not None else 0.0
-                unit_price_item = QTableWidgetItem(f"€ {effective_unit_price:.2f}")
-                unit_price_item.setTextAlignment(Qt.AlignRight | Qt.AlignVCenter)
-                self.products_table.setItem(row_idx, 4, unit_price_item)
-                
-                total_price_calculated_val = prod_link_data.get('total_price_calculated')
-                if total_price_calculated_val is None:
-                    total_price_calculated_val = 0.0
-                total_price_item = QTableWidgetItem(f"€ {float(total_price_calculated_val):.2f}")
-                total_price_item.setTextAlignment(Qt.AlignRight | Qt.AlignVCenter)
-                self.products_table.setItem(row_idx, 5, total_price_item)
-                
-            self.products_table.resizeColumnsToContents()
-            
-        except Exception as e:
-            QMessageBox.warning(self, self.tr("Erreur DB"), self.tr("Erreur de chargement des produits:\n{0}").format(str(e)))
 
 class EditClientDialog(QDialog):
     def __init__(self, client_info, config, parent=None):
@@ -4042,105 +2699,6 @@ class DocumentManager(QMainWindow):
     def closeEvent(self, event): 
         save_config(self.config) 
         super().closeEvent(event)
-
-class SettingsDialog(QDialog):
-    def __init__(self, current_config, parent=None): 
-        super().__init__(parent)
-        self.setWindowTitle(self.tr("Paramètres de l'Application")); self.setMinimumSize(500, 400)
-        self.current_config_data = current_config 
-        self.setup_ui_settings() 
-        
-    def setup_ui_settings(self): 
-        layout = QVBoxLayout(self); tabs_widget = QTabWidget(); layout.addWidget(tabs_widget) 
-        
-        general_tab_widget = QWidget(); general_form_layout = QFormLayout(general_tab_widget) 
-        self.templates_dir_input = QLineEdit(self.current_config_data["templates_dir"]) 
-        templates_browse_btn = QPushButton(self.tr("Parcourir...")); templates_browse_btn.clicked.connect(lambda: self.browse_directory_for_input(self.templates_dir_input, self.tr("Sélectionner dossier modèles")))
-        templates_dir_layout = QHBoxLayout(); templates_dir_layout.addWidget(self.templates_dir_input); templates_dir_layout.addWidget(templates_browse_btn) 
-        general_form_layout.addRow(self.tr("Dossier des Modèles:"), templates_dir_layout)
-        
-        self.clients_dir_input = QLineEdit(self.current_config_data["clients_dir"]) 
-        clients_browse_btn = QPushButton(self.tr("Parcourir...")); clients_browse_btn.clicked.connect(lambda: self.browse_directory_for_input(self.clients_dir_input, self.tr("Sélectionner dossier clients")))
-        clients_dir_layout = QHBoxLayout(); clients_dir_layout.addWidget(self.clients_dir_input); clients_dir_layout.addWidget(clients_browse_btn) 
-        general_form_layout.addRow(self.tr("Dossier des Clients:"), clients_dir_layout)
-
-        self.interface_lang_combo = QComboBox()
-        # Supported languages: French, English, Arabic, Turkish, Portuguese
-        self.lang_display_to_code = {
-            self.tr("Français (fr)"): "fr",
-            self.tr("English (en)"): "en",
-            self.tr("العربية (ar)"): "ar",
-            self.tr("Türkçe (tr)"): "tr",
-            self.tr("Português (pt)"): "pt"
-        }
-        self.interface_lang_combo.addItems(list(self.lang_display_to_code.keys()))
-
-        current_lang_code = self.current_config_data.get("language", "fr") # Default to 'fr' if not set
-        
-        # Find the display text for the current language code to set the combo box
-        # This creates a reverse mapping from code to display text for initial setting
-        code_to_display_text = {code: display for display, code in self.lang_display_to_code.items()}
-
-        current_display_text = code_to_display_text.get(current_lang_code)
-        if current_display_text:
-            self.interface_lang_combo.setCurrentText(current_display_text)
-        else:
-            # Fallback if current_lang_code from config isn't in our map (e.g. old config value)
-            # Set to French display text by default
-            french_display_text = code_to_display_text.get("fr", list(self.lang_display_to_code.keys())[0]) # Get "Français (fr)" or first item
-            self.interface_lang_combo.setCurrentText(french_display_text)
-
-        general_form_layout.addRow(self.tr("Langue Interface (redémarrage requis):"), self.interface_lang_combo)
-        
-        self.reminder_days_spinbox = QSpinBox(); self.reminder_days_spinbox.setRange(1, 365) 
-        self.reminder_days_spinbox.setValue(self.current_config_data.get("default_reminder_days", 30))
-        general_form_layout.addRow(self.tr("Jours avant rappel client ancien:"), self.reminder_days_spinbox)
-        tabs_widget.addTab(general_tab_widget, self.tr("Général"))
-        
-        email_tab_widget = QWidget(); email_form_layout = QFormLayout(email_tab_widget) 
-        self.smtp_server_input_field = QLineEdit(self.current_config_data.get("smtp_server", "")) 
-        email_form_layout.addRow(self.tr("Serveur SMTP:"), self.smtp_server_input_field)
-        self.smtp_port_spinbox = QSpinBox(); self.smtp_port_spinbox.setRange(1, 65535) 
-        self.smtp_port_spinbox.setValue(self.current_config_data.get("smtp_port", 587))
-        email_form_layout.addRow(self.tr("Port SMTP:"), self.smtp_port_spinbox)
-        self.smtp_user_input_field = QLineEdit(self.current_config_data.get("smtp_user", "")) 
-        email_form_layout.addRow(self.tr("Utilisateur SMTP:"), self.smtp_user_input_field)
-        self.smtp_pass_input_field = QLineEdit(self.current_config_data.get("smtp_password", "")) 
-        self.smtp_pass_input_field.setEchoMode(QLineEdit.Password)
-        email_form_layout.addRow(self.tr("Mot de passe SMTP:"), self.smtp_pass_input_field)
-        tabs_widget.addTab(email_tab_widget, self.tr("Email"))
-
-        # Company Details Tab
-        self.company_tab = CompanyTabWidget(self) # Create instance of CompanyTabWidget
-        tabs_widget.addTab(self.company_tab, self.tr("Company Details")) # Add it as a tab
-        
-        dialog_button_box = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        ok_settings_button = dialog_button_box.button(QDialogButtonBox.Ok)
-        ok_settings_button.setText(self.tr("OK"))
-        ok_settings_button.setObjectName("primaryButton")
-
-        cancel_settings_button = dialog_button_box.button(QDialogButtonBox.Cancel)
-        cancel_settings_button.setText(self.tr("Annuler"))
-        # Standard cancel button styling will apply from global stylesheet
-
-        dialog_button_box.accepted.connect(self.accept); dialog_button_box.rejected.connect(self.reject)
-        layout.addWidget(dialog_button_box)
-        
-    def browse_directory_for_input(self, line_edit_target, dialog_title):
-        dir_path = QFileDialog.getExistingDirectory(self, dialog_title, line_edit_target.text())
-        if dir_path: line_edit_target.setText(dir_path)
-            
-    def get_config(self):
-        # Get the current display text from combo box and map it back to language code
-        selected_lang_display_text = self.interface_lang_combo.currentText()
-        language_code = self.lang_display_to_code.get(selected_lang_display_text, "fr") # Default to fr if something goes wrong
-        return {
-            "templates_dir": self.templates_dir_input.text(), "clients_dir": self.clients_dir_input.text(),
-            "language": language_code,
-            "default_reminder_days": self.reminder_days_spinbox.value(),
-            "smtp_server": self.smtp_server_input_field.text(), "smtp_port": self.smtp_port_spinbox.value(),
-            "smtp_user": self.smtp_user_input_field.text(), "smtp_password": self.smtp_pass_input_field.text()
-        }
 
 def main():
     if hasattr(Qt, 'AA_EnableHighDpiScaling'): QApplication.setAttribute(Qt.AA_EnableHighDpiScaling, True)
